[RELEASE] - 1.2.6 (#1433)
Bumping version and moving routes to be one file
