+++ containers/ecr-viewer/.env
@@ -1,5 +1,5 @@
 DATABASE_URL=postgres://postgres:pw@localhost:5432/ecr_viewer_db
 AWS_REGION=us-east-1
 ECR_BUCKET_NAME=ecr-viewer-files
-NEXT_PUBLIC_SOURCE=postgres
+SOURCE=postgres
 NEXT_TELEMETRY_DISABLED=1
