+++ containers/ecr-viewer/src/app/api/fhir-data/route.ts
@@ -1,8 +1,23 @@
 import { NextRequest, NextResponse } from "next/server";
 import pgPromise from "pg-promise";
-import { loadYamlConfig } from "../utils";
+import { S3Client, GetObjectCommand } from "@aws-sdk/client-s3";
+import { loadYamlConfig, streamToJson } from "../utils";
+
+const S3_SOURCE = "s3";
+const POSTGRES_SOURCE = "postgres";
+const s3Client = new S3Client({ region: process.env.AWS_REGION });
 
 export async function GET(request: NextRequest) {
+  if (process.env.SOURCE === S3_SOURCE) {
+    return get_s3(request);
+  } else if (process.env.SOURCE === POSTGRES_SOURCE) {
+    return await get_postgres(request);
+  } else {
+    return NextResponse.json({ message: "Invalid source" }, { status: 500 });
+  }
+}
+
+const get_postgres = async (request: NextRequest) => {
   const params = request.nextUrl.searchParams;
   const ecr_id = params.get("id") ? params.get("id") : null;
   const db_url = process.env.DATABASE_URL || "";
@@ -26,4 +41,29 @@ export async function GET(request: NextRequest) {
     console.error("Error fetching data:", error);
     return NextResponse.json({ message: error.message }, { status: 500 });
   }
-}
+};
+
+const get_s3 = async (request: NextRequest) => {
+  const params = request.nextUrl.searchParams;
+  const ecr_id = params.get("id");
+  const bucketName = process.env.ECR_BUCKET_NAME;
+  const objectKey = `${ecr_id}.json`; // This could also come from the request, e.g., req.query.key
+
+  try {
+    const command = new GetObjectCommand({
+      Bucket: bucketName,
+      Key: objectKey,
+    });
+
+    const { Body } = await s3Client.send(command);
+    const content = await streamToJson(Body);
+
+    return NextResponse.json(
+      { fhirBundle: content, fhirPathMappings: loadYamlConfig() },
+      { status: 200 },
+    );
+  } catch (error: any) {
+    console.error("S3 GetObject error:", error);
+    return NextResponse.json({ message: error.message }, { status: 500 });
+  }
+};
