+++ containers/ecr-viewer/src/app/view-data/page.tsx
@@ -10,16 +10,6 @@ import { processSnomedCode } from "./service";
 
 // string constants to match with possible .env values
 const basePath = process.env.NODE_ENV === "production" ? "/ecr-viewer" : "";
-const S3_SOURCE = "s3";
-const POSTGRES_SOURCE = "postgres";
-
-const assignApiPath = () => {
-  if (process.env.NEXT_PUBLIC_SOURCE === S3_SOURCE) {
-    return "s3";
-  } else if (process.env.NEXT_PUBLIC_SOURCE === POSTGRES_SOURCE) {
-    return "fhir-data";
-  }
-};
 
 const ECRViewerPage = () => {
   const [fhirBundle, setFhirBundle] = useState<Bundle>();
@@ -27,7 +17,6 @@ const ECRViewerPage = () => {
   const [errors, setErrors] = useState<Error | unknown>(null);
   const searchParams = useSearchParams();
   const fhirId = searchParams.get("id") ?? "";
-  const apiPath = assignApiPath();
   const snomedCode = searchParams.get("snomed-code") ?? "";
 
   type ApiResponse = {
@@ -40,7 +29,7 @@ const ECRViewerPage = () => {
 
     const fetchData = async () => {
       try {
-        const response = await fetch(`${basePath}/api/${apiPath}?id=${fhirId}`);
+        const response = await fetch(`${basePath}/api/fhir-data?id=${fhirId}`);
         if (!response.ok) {
           const errorData = response.statusText;
           throw new Error(errorData || "Internal Server Error");
