+++ phdi/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.2.5"
+__version__ = "1.2.6"
