+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "phdi"
-version = "v1.2.5"
+version = "v1.2.6"
 description = "Public health data infrastructure Building Blocks is a library to help public health departments work with their data"
 authors = [
   "Kenneth Chow <kenneth@skylight.digital>",
